pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "role-radar"
version = "1.0.0"
description = "CLI tool to find Product Manager roles at top AI companies and VC-backed startups"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Role Radar", email = "roleradar@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "tenacity>=8.2.0",
    "pypdf>=3.15.0",
    "python-docx>=1.0.0",
    "rich>=13.0.0",
    "structlog>=23.0.0",
    "sendgrid>=6.10.0",
    "urllib3>=2.0.0",
    "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "responses>=0.23.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
openai = [
    "openai>=1.0.0",
]

[project.scripts]
role-radar = "role_radar.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
role_radar = ["templates/*.html", "templates/*.txt", "data/*.yaml", "data/*.csv", "web/templates/*.html", "web/static/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=role_radar --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
